refactor(accelerator): extract server::tls module (start_https) from server.rs (Q2 2/n)
Second module of the server.rs split: start_https (the HTTPS listener for optional
Safari support, with its own AddrInUse retry) moves verbatim into
src/server/tls.rs. server.rs gets `mod tls; pub use tls::start_https;` (so
server::start_https stays the public path for main.rs + commands.rs) and drops the
now-unused TlsAcceptor import. tls.rs imports router/AppState/HTTPS_PORT +
bind::bind_with_retry from the parent.

Behavior-preserving: 127 lib tests + clippy -D warnings + bin (re-export resolves)
+ headless all green. Stacked on the bind extraction (Q2 1/n, #309).
